code/requirements.txt
boto3==1.35.74
botocore==1.35.74
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
jmespath==1.0.1
numpy==2.1.3
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
s3transfer==0.10.4
six==1.16.0
tzdata==2024.2
urllib3==2.2.3